Deleted unused files and moved the URL handlers into their own file. Changed login logic slightly
