+++ assets/templates/404.html
@@ -20,7 +20,7 @@
         <img src="/assets/static/img/404.jpg" class="img-responsive inline-block" alt="kev" />
       </div>
       <div class="row">
-        <h2>Your page could not be found but please enjoy this picture of Kevin</h2>
+        <h2>Your page could not be found but please enjoy this picture of Kevin having fun with all his friends</h2>
       </div>
     </div>
