--- assets/templates/loggedInPage.html.tmpl
@@ -1,1 +1,0 @@
-<h1>You are logged in!</h1>
