+++ url_handlers.go
@@ -1,0 +1,160 @@
+package main
+
+import (
+    "log"
+    "net/http"
+    "html/template"
+    "golang.org/x/crypto/bcrypt"
+    "github.com/gorilla/sessions"
+    "math/rand"
+    _ "github.com/jinzhu/gorm/dialects/postgres"
+  )
+
+func MakeRandAlphaNumericStr(n int) string {
+  var letters = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789")
+  b := make([]rune, n)
+  for i := range b {
+      b[i] = letters[rand.Intn(len(letters))]
+  }
+  return string(b)
+}
+
+func Handle500Error(w http.ResponseWriter, err error) {
+  http.Error(w, http.StatusText(500), 500)
+  log.Fatal(err)
+}
+
+func GetSession(w http.ResponseWriter, r *http.Request) *sessions.Session {
+  session, err := sessionStore.Get(r, "buzzer-session")
+  if err != nil {
+    Handle500Error(w, err)
+  }
+  return session
+}
+
+func RenderTemplate(w http.ResponseWriter, template_name string, template_params map[string] interface{}) {
+  t, err := template.ParseFiles(template_name)
+  if err != nil{
+    Handle500Error(w, err)
+  }
+  t.Execute(w, template_params)
+}
+
+func AddFlashToSession(w http.ResponseWriter, r *http.Request, flash string, session *sessions.Session) {
+  session.AddFlash(flash)
+  session.Save(r, w)
+}
+
+func IsUserLoggedIn(session *sessions.Session) bool {
+  username, found := session.Values["username"]
+  return found && username != ""
+}
+
+func LoginHandler(w http.ResponseWriter, r *http.Request) {
+  log.SetPrefix("[LoginURLHandler] ")
+  session := GetSession(w, r)
+  if IsUserLoggedIn(session) {
+    http.Redirect(w, r, "/wait_list", 302)
+    return
+  }
+  if r.Method == "POST" {
+    log.Println("in post")
+    username := r.FormValue("username")
+    password := r.FormValue("password")
+
+    if username == "" || password == "" {
+      AddFlashToSession(w, r, "Missing form field", session)
+    } else {
+      var user User;
+      db.First(&user, "Username = ?", username)
+      if (user != (User{})) {
+        passSalt := user.PassSalt
+        if (bcrypt.CompareHashAndPassword([]byte(user.Password), []byte(password+passSalt)) == nil) {
+            session.Values["username"] = username
+            session.Save(r, w)
+            http.Redirect(w, r, "/wait_list", 302)
+            return
+        }
+      }
+      AddFlashToSession(w, r, "Username or password is incorrect", session)
+    }
+  }
+  template_data := map[string]interface{}{}
+  if flashes := session.Flashes(); len(flashes) > 0 {
+      template_data["failure_message"] = flashes[0]
+  }
+  session.Save(r, w)
+  RenderTemplate(w, "assets/templates/login.html.tmpl", template_data)
+}
+
+func WaitListHandler(w http.ResponseWriter, r *http.Request) {
+  log.SetPrefix("[WaitListHandler] ")
+  if !IsUserLoggedIn(GetSession(w, r)) {
+    http.Redirect(w, r, "/login", 302)
+    return
+  }
+  RenderTemplate(w, "assets/templates/waitlist.html.tmpl", nil)
+}
+
+
+func RootHandler(w http.ResponseWriter, r *http.Request) {
+  log.SetPrefix("[RootHandler] ")
+  RenderTemplate(w, "assets/templates/login.html.tmpl", nil)
+}
+
+func AddUserHandler(w http.ResponseWriter, r *http.Request) {
+  log.SetPrefix("[AddUserHandler] ")
+  session := GetSession(w, r)
+  if r.Method == "POST" {
+    username := r.FormValue("username")
+    password := r.FormValue("password")
+    restaurantName := r.FormValue("restaurant_name")
+    if username != "" && password != "" && restaurantName != "" {
+
+      //salt and hash the password
+      passSalt := MakeRandAlphaNumericStr(50)
+      hashedPassword, err := bcrypt.GenerateFromPassword([]byte(password+passSalt), bcrypt.DefaultCost)
+      if err != nil {
+        log.Fatal(err)
+      }
+
+      var restaurant Restaurant;
+      db.First(&restaurant, "name = ?", restaurantName)
+
+      //make a restaurant if there isn't one
+      if restaurant == (Restaurant{}) {
+        restaurant = Restaurant{Name: restaurantName}
+        db.NewRecord(restaurant)
+        db.Create(&restaurant)
+      }
+
+      var user User
+      db.First(&user, "username = ?", username)
+
+      if user != (User{}) {
+        AddFlashToSession(w, r, "Username already exists", session)
+      } else {
+        //add the user
+        user = User{RestaurantID: restaurant.ID, Username: username, Password: string(hashedPassword), PassSalt: passSalt}
+        db.NewRecord(user)
+        db.Create(&user)
+        AddFlashToSession(w, r, "User successfully added", session)
+      }
+    } else {
+      AddFlashToSession(w, r, "Could not add user. Did you forget a field?", session)
+    }
+    http.Redirect(w, r, "/add_user", 302)
+  } else {
+    template_data := map[string]interface{}{}
+    if flashes := session.Flashes(); len(flashes) > 0 {
+      template_data["flash"] = flashes[0]
+    }
+    session.Save(r, w)
+    RenderTemplate(w, "assets/templates/adduser.html.tmpl", template_data)
+  }
+}
+
+func NotFoundHandler(w http.ResponseWriter, r *http.Request) {
+  log.SetPrefix("[NotFoundHandler] ")
+  RenderTemplate(w, "assets/templates/404.html", nil)
+}
